Extract Json assertion builder for reuse - fixes #86
